fix(warnings): Fix C4464 and C5219 compiler warnings
C5219 (implicit conversion to double):
- benchmark.hpp: Add explicit static_cast<double>(mft_bytes)
- main_dialog.hpp: Add explicit static_cast<double>(item_count)

C4464 (relative include path contains '..'):
- Add '.' and 'src' to AdditionalIncludeDirectories in all 4 build configs
- Update 98 include statements across 26 files to use absolute paths
  from src/ directory instead of relative '../' paths

The '.' in include path allows finding stdafx.h, resource.h, and CLI11.hpp
which are in the project root directory.

+++ UltraFastFileSearch-code/src/cli/benchmark.hpp
@@ -178,7 +178,7 @@ inline int benchmark_index_build(char drive_letter, std::ostream& OS)
 
     // Calculate throughput
     double mb_per_sec = (seconds > 0)
-        ? (mft_bytes / (1024.0 * 1024.0)) / seconds
+        ? (static_cast<double>(mft_bytes) / (1024.0 * 1024.0)) / seconds
         : 0;
     double records_per_sec = (seconds > 0)
         ? static_cast<double>(total_records) / seconds
